Quick norminette fix + reorganization of shapes' helper functions
