+++ srcs/shapes/cylinder_utils.c
@@ -12,20 +12,23 @@ void	transform_cy(t_obj *shape, t_matrix transformation)
 			inverse(transformation, 4));
 }
 
-double	discriminant_cyl(t_obj *cyl, double *a, double *b)
+t_tup	local_normal_at_cy(t_obj *cyl, t_tup point)
 {
-	double	c;
-	double	disc;
+	double		dist;
+	t_cyl		*cyll;
+	t_tup		normal;
 
-	*a = pow(cyl->saved_ray.direction.x, 2) \
-		+ pow(cyl->saved_ray.direction.z, 2);
-	if (fabs(*a) < DBL_EPSILON)
-		return (-1);
-	*b = 2 * cyl->saved_ray.point.x * cyl->saved_ray.direction.x \
-		+ 2 * cyl->saved_ray.point.z * cyl->saved_ray.direction.z;
-	c = pow(cyl->saved_ray.point.x, 2) + pow(cyl->saved_ray.point.z, 2) - 1;
-	disc = pow(*b, 2) - 4 * *a * c;
-	return (disc);
+	cyll = (t_cyl *) cyl;
+	dist = pow(point.x, 2) + pow(point.z, 2);
+	if (dist < 1 && point.y >= cyll->max - EPSILON)
+		normal = vector(0, 1, 0);
+	else if (dist < 1 && point.y <= cyll->min + EPSILON)
+		normal = vector(0, -1, 0);
+	else
+		normal = vector_norm(vector(point.x, 0, point.z));
+	if (cyl->material.pattern.noise)
+		normal = perturb_normal(cyl, point, normal, cyl->material.pattern);
+	return (normal);
 }
 
 t_inter	*local_intersect_cy(t_ray r, t_obj *cyl)
@@ -56,7 +59,6 @@ t_inter	*local_intersect_cy(t_ray r, t_obj *cyl)
 	return (ret);
 }
 
-
 t_cyl	*cylinder(int i)
 {
 	t_cyl	*cyl;
@@ -64,21 +66,12 @@ t_cyl	*cylinder(int i)
 	cyl = malloc(sizeof(t_cyl));
 	if (!cyl)
 		return (NULL);
+	cyl_set_up(cyl, i);
 	cyl->diameter = 2;
 	cyl->height = 1;
-	cyl->shape.get_name = get_name_cy;
-	cyl->shape.volume = volume_cy;
-	cyl->shape.destroy = cyl_destroy;
-	cyl->shape.transform = transform_cy;
-	cyl->shape.local_intersect = local_intersect_cy;
 	cyl->shape.material = mat_default();
 	cyl->shape.point = point(0, 0, 0);
 	cyl->shape.dir_vector = vector(0, 0, 1);
-	cyl->shape.next = NULL;
-	cyl->shape.transformation = identity();
-	cyl->shape.inverse_transformation = identity();
-	cyl->shape.id = i;
-	cyl->closed = true;
 	cyl->max = DBL_MAX;
 	cyl->min = DBL_MIN;
 	return (cyl);
