+++ srcs/shapes/sphere_utils.c
@@ -12,7 +12,8 @@ t_tup	local_normal_at_s(t_obj *sphere, t_tup point)
 
 	normal = normal_at(sphere, point);
 	if (sphere->material.pattern.noise)
-		normal = perturb_normal(sphere, point, normal, sphere->material.pattern);
+		normal = perturb_normal(sphere, point, normal,
+				sphere->material.pattern);
 	return (normal);
 }
 
@@ -23,20 +24,11 @@ t_sphere	*sphere(int i)
 	s = malloc(sizeof(t_sphere));
 	if (!s)
 		return (NULL);
+	sphere_set_up(s, i);
 	s->diameter = 2;
-	s->shape.get_name = get_name_s;
-	s->shape.volume = volume_s;
-	s->shape.destroy = sphere_destroy;
-	s->shape.transform = transform_s;
-	s->shape.local_intersect = local_intersect_s;
-	s->shape.local_normal_at = local_normal_at_s;
 	s->shape.point = point(0, 0, 0);
 	s->shape.dir_vector = vector(0, 0, 0);
 	s->shape.material = mat_default();
-	s->shape.next = NULL;
-	s->shape.id = i;
-	s->shape.transformation = identity();
-	s->shape.inverse_transformation = identity();
 	s->shape.transform((t_obj *) s,
 		scaling_matrix(s->diameter / 2, s->diameter / 2, s->diameter / 2));
 	return (s);
@@ -49,23 +41,28 @@ t_sphere	*glass_sphere(int i)
 	s = malloc(sizeof(t_sphere));
 	if (!s)
 		return (NULL);
+	sphere_set_up(s, i);
 	s->diameter = 2;
-	s->shape.get_name = get_name_s;
-	s->shape.volume = volume_s;
-	s->shape.destroy = sphere_destroy;
-	s->shape.transform = transform_s;
-	s->shape.local_intersect = local_intersect_s;
-	s->shape.local_normal_at = local_normal_at_s;
 	s->shape.point = point(0, 0, 0);
 	s->shape.dir_vector = vector(0, 0, 0);
 	s->shape.material = mat_default();
 	s->shape.material.refractive_index = 1.5;
 	s->shape.material.transparency = 1;
-	s->shape.next = NULL;
-	s->shape.id = i;
-	s->shape.transformation = identity();
-	s->shape.inverse_transformation = identity();
 	s->shape.transform((t_obj *) s,
 		scaling_matrix(s->diameter / 2, s->diameter / 2, s->diameter / 2));
 	return (s);
+}
+
+void	sphere_set_up(t_sphere *sphere, int i)
+{
+	sphere->shape.id = i;
+	sphere->shape.get_name = get_name_s;
+	sphere->shape.volume = volume_s;
+	sphere->shape.destroy = sphere_destroy;
+	sphere->shape.transform = transform_s;
+	sphere->shape.local_intersect = local_intersect_s;
+	sphere->shape.local_normal_at = local_normal_at_s;
+	sphere->shape.next = NULL;
+	sphere->shape.transformation = identity();
+	sphere->shape.inverse_transformation = identity();
 }
